+++ opal/core/pathway/pathways.py
@@ -71,6 +71,7 @@ class Pathway(discoverable.DiscoverableFeature):
             activepattern=kwargs.get('activepattern', kls.get_absolute_url()),
             icon=kwargs.get('icon', kls.get_icon()),
             display=kwargs.get('display', kls.get_display_name()),
+            index=kwargs.get('index', '')
         )
 
     def get_pathway_service(self, is_modal):