+++ changelog.md
@@ -113,13 +113,20 @@ including the `reopen_episode_modal.html` template and the url/view at `template
 
 * `__unicode__` model methods have been renamed `__str__`
 
+* Adds an index argument to `PatientList.as_menuitem()` and `Pathway.as_menuitem()`
+
+* Adds a `get_absolute_url()` method to `Patient` and `Episode
+
+
 #### Updates to the Dependency Graph
 
 * Django: 1.10.8 -> 2.0.9
 * Django Rest Framework: 3.4.7 -> 3.7.4
 * Django Reversion: 1.10.2 -> 3.0.1
 * Letter: 0.4.1 -> 0.5
-
+* Requests: 2.18.4 -> 2.20.1
+* Psycopg2: 2.7 -> 2.7.6.1
+* Python Dateutil: 2.4.2 -> 2.7.5
 
 ### 0.12.0 (Major Release)